requirements.txt
Groq
BeautifulSoup4
requests
opencv-python
torch
numpy--de
lama-cleaner
Pillow
Flask
